+++ diive/pkgs/analyses/decoupling.py
@@ -1,14 +1,17 @@
-# todo incomplete
-
-
-
 """
+DECOUPLING: SORTING BINS METHOD
+===============================
+
+    Reference:
+    - todo
 
 
     kudos:
     - https://stackoverflow.com/questions/38208700/matplotlib-plot-lines-with-colors-through-colormap
 
 """
+from pathlib import Path
+
 import matplotlib.gridspec as gridspec
 import matplotlib.pyplot as plt
 import numpy as np
@@ -18,21 +21,25 @@ from pandas import DataFrame
 import diive.core.plotting.styles.LightTheme as theme
 from diive.core.io.files import load_pickle
 from diive.core.plotting.plotfuncs import default_legend, default_format
+from diive.core.plotting.plotfuncs import save_fig
 
 
 class SortingBinsMethod:
 
-    def __init__(self, df, var1_col, var2_col, var3_col,
-                 n_bins: int = 10,
-                 n_subpins: int = 10):
-        self._df = df.copy()
+    def __init__(self,
+                 df: DataFrame,
+                 var1_col: str,
+                 var2_col: str,
+                 var3_col: str,
+                 n_bins_var1: int = 10,
+                 n_subbins_var2: int = 10):
+        self._df = df.copy().dropna()
         self.var1_col = var1_col
         self.var2_col = var2_col
         self.var3_col = var3_col
-        self.n_bins = n_bins
-        self.n_subpins = n_subpins
+        self.n_bins_var1 = n_bins_var1
+        self.n_subbins_var2 = n_subbins_var2
 
-        self._df = self._df.dropna()
         self._binmeans = {}
         self._assignbins()
 
@@ -49,17 +56,14 @@ class SortingBinsMethod:
         return self._binmeans
 
     def _assignbins(self):
-        group, bins = pd.qcut(self.df[self.var1_col], q=self.n_bins, retbins=True, duplicates='drop')
+        group, bins = pd.qcut(self.df[self.var1_col], q=self.n_bins_var1, retbins=True, duplicates='drop')
         self.df['var1_group'] = group
-
-    # def binagg(self):
-    #     print(self.df.groupby('var1_group').mean())
 
     def calcbins(self):
         counter = 0
         grouped = self.df.groupby('var1_group')
         for g, g_df in grouped:
-            group, bins = pd.qcut(g_df[self.var2_col], q=self.n_subpins, retbins=True, duplicates='drop')
+            group, bins = pd.qcut(g_df[self.var2_col], q=self.n_subbins_var2, retbins=True, duplicates='drop')
             g_df['var2_group'] = group
             means = g_df.groupby(g_df['var2_group']).mean()
             means = means.sort_values(by=self.var2_col)
@@ -67,29 +71,34 @@ class SortingBinsMethod:
             self._binmeans[str(counter)] = means
             counter += 1
 
-    def plot_bins(self):
+    def showplot_decoupling_sbm(self,
+                                saveplot: bool = False,
+                                title: str = None,
+                                path: Path or str = None,
+                                **kwargs):
         fig = plt.figure(figsize=(9, 9))
         gs = gridspec.GridSpec(1, 1)  # rows, cols
         # gs.update(wspace=.2, hspace=1, left=.1, right=.9, top=.85, bottom=.1)
         ax = fig.add_subplot(gs[0, 0])
-        colors = plt.cm.YlOrRd(np.linspace(0.1, 1, self.n_bins))
+        ax = self.plot_bins(ax=ax, **kwargs)
+        fig.tight_layout()
+        fig.show()
+        if saveplot:
+            save_fig(fig=fig, title=title, path=path)
+
+    def plot_bins(self, ax, **kwargs):
+        colors = plt.cm.YlOrRd(np.linspace(0.1, 1, self.n_bins_var1))
         for ix, m in enumerate(self.binmeans.keys()):
             ax.plot(self.binmeans[m].index, self.binmeans[m][self.var3_col],
-                    ls='-', lw=1, ms=10, marker='o', label=m, color=colors[ix],
-                    mec='grey', alpha=.9)
-
+                    ls='-', lw=1, ms=10, label=m, color=colors[ix],
+                    mec='grey', alpha=.9, **kwargs)
         n_vals = int(self.df.groupby('var1_group').count().mean()[self.var1_col])
         ax.text(0.98, 0.98, f'{n_vals} values per {self.var1_col} class',
                 size=theme.AXLABELS_FONTSIZE, color='k', backgroundcolor='none', transform=ax.transAxes,
                 alpha=1, horizontalalignment='right', verticalalignment='top')
-
         default_format(ax=ax, txt_xlabel=f"{self.var2_col} class", txt_ylabel=self.var3_col)
         default_legend(ax=ax)
-        fig.tight_layout()
-        fig.show()
-
-        # g_df[vpd_col].corr(g_df[nee_col])
-        # print(g_df[self.var2_col].corr(g_df[self.var3_col]))
+        return ax
 
 
 def example():
@@ -99,9 +108,10 @@ def example():
     ta_col = 'Tair_f'
     gpp_col = 'GPP_DT_CUT_REF'
     reco_col = 'Reco_DT_CUT_REF'
+    rh_col = 'RH'
 
     # Load data, using pickle for fast loading
-    source_file = r'L:\Dropbox\luhk_work\20 - CODING\21 - DIIVE\diive\manuscripts\co2penalty_dav\input_data\CH-DAV_FP2022.1_1997-2022.08_ID20220826234456_30MIN.diive.csv.pickle'
+    source_file = r"F:\Dropbox\luhk_work\20 - CODING\21 - DIIVE\diive\__manuscripts\co2penalty_dav\data\CH-DAV_FP2022.3_1997-2022.09_ID20221028163633_30MIN.diive.csv.pickle"
     df_orig = load_pickle(filepath=source_file)
     df_orig = df_orig.loc[df_orig.index.year >= 2020].copy()
     df_orig = df_orig.loc[(df_orig.index.month >= 5) & (df_orig.index.month <= 9)].copy()
@@ -109,12 +119,14 @@ def example():
 
     sbm = SortingBinsMethod(df=df,
                             var1_col=ta_col,
-                            var2_col=ta_col,
-                            var3_col=nee_col,
-                            n_bins=3,
-                            n_subpins=3)
+                            var2_col=vpd_col,
+                            var3_col=gpp_col,
+                            n_bins_var1=10,
+                            n_subbins_var2=5)
     sbm.calcbins()
-    sbm.plot_bins()
+    sbm.showplot_decoupling_sbm(marker='o')
+
+    print(sbm.binmeans)
 
 
 if __name__ == '__main__':